Fix and clean up
Former-commit-id: 346460fc0474b25365abb4c08b54d0e663e40773 [formerly b83fa8d3ffdaa448e9680b92a05c85b48051e75f] [formerly e9c673478f851201f45ba8dfa9bfa93e5ec25350 [formerly 22ff9ddcf1fe3a78278833897480db2826b249bd] [formerly 8ed4cbd9e3c05f6d7ccd8e03e323823dd030cecb] [formerly c7e426f5b61bae5ecab4e51900ff27fc2beca540 [formerly 92ba4d6e5116d5b29aeb53c73eca25e9c7ca18e0]] [formerly e9c673478f851201f45ba8dfa9bfa93e5ec25350 [formerly 22ff9ddcf1fe3a78278833897480db2826b249bd] [formerly 8ed4cbd9e3c05f6d7ccd8e03e323823dd030cecb] [formerly c7e426f5b61bae5ecab4e51900ff27fc2beca540 [formerly 92ba4d6e5116d5b29aeb53c73eca25e9c7ca18e0]] [formerly 8e656a8932ea6b9dea33bf6d464cc41420ee9961 [formerly b55401176dd70f7a7721ce94160b78767b7c8d60]]]]
Former-commit-id: 1046e3273e1d908d6c37798f31dbadd3e7235b0f
Former-commit-id: 0fd9826f36758da7d57988032708fc278f87ce61
Former-commit-id: 163849ceeea24fcb97faf4359db7b54d80637a19
Former-commit-id: 421cfc4e894321220a53a8db2f20f1bd86894da4
Former-commit-id: 604b659868defcfb3af8d86b3e888e9961b49564 [formerly 7c8dcfe2726e838b587a7551a53d757418e34f56]
Former-commit-id: 5b60de8bb1d936d7023c6596e5c7ebd244347cdd
Former-commit-id: 68258e9bfddfb7ccb0addb5014061ada6231327a

+++ Source/RawParser/Model/Effect/Demosaic.cs
@@ -58,14 +58,14 @@ namespace RawEditor.Effect
 
         private static void Deflate(RawImage image, ushort[] deflated)
         {
-            int height = image.colorFilter.Size.height;
-            int width = image.colorFilter.Size.width;
+            uint height = image.colorFilter.Size.height;
+            uint width = image.colorFilter.Size.width;
             Parallel.For(0, image.raw.dim.height, row =>
             {
-                int realRow = row + image.raw.offset.height;
+                long realRow = row + image.raw.offset.height;
                 for (int col = 0; col < image.raw.dim.width; col++)
                 {
-                    int realCol = col + image.raw.offset.width;
+                    long realCol = col + image.raw.offset.width;
                     int pixeltype = (int)image.colorFilter.cfa[((row % height) * width) + col % width];
 
                     deflated[(((row * image.raw.dim.width) + col) * 3) + pixeltype] = image[realRow, realCol];
@@ -75,16 +75,16 @@ namespace RawEditor.Effect
 
         private static void FujiBilinear(RawImage image, ushort[] deflated)
         {
-            int height = image.colorFilter.Size.height;
-            int width = image.colorFilter.Size.width;
+            uint height = image.colorFilter.Size.height;
+            uint width = image.colorFilter.Size.width;
             //first deflate
             Deflate(image, deflated);
             Parallel.For(0, image.raw.dim.height, row =>
             {
-                int realRow = row + image.raw.offset.height;
+                long realRow = row + image.raw.offset.height;
                 for (int col = 0; col < image.raw.dim.width; col++)
                 {
-                    int realCol = col + image.raw.offset.width;
+                    long realCol = col + image.raw.offset.width;
                     CFAColor pixeltype = image.colorFilter.cfa[((row % height) * width) + col % width];
                     if (pixeltype == CFAColor.GREEN)
                     {
@@ -117,10 +117,10 @@ namespace RawEditor.Effect
         {
             Parallel.For(0, image.raw.dim.height, row =>
             {
-                int realRow = row + image.raw.offset.height;
+                long realRow = row + image.raw.offset.height;
                 for (int col = 0; col < image.raw.dim.width; col++)
                 {
-                    int realCol = col + image.raw.offset.width;
+                    long realCol = col + image.raw.offset.width;
                     CFAColor pixeltype = image.colorFilter.cfa[((row % 2) * 2) + col % 2];
                     if (pixeltype == CFAColor.GREEN)
                     {
