+++ Source/RawParser/Model/Parser/Decoder/JPGDecoder.cs
@@ -59,19 +59,19 @@ namespace RawNet.Decoder
             using (IMemoryBufferReference reference = buffer.CreateReference())
             {
                 BitmapPlaneDescription bufferLayout = buffer.GetPlaneDescription(0);
-                rawImage.raw.dim = new Point2D(bufferLayout.Width, bufferLayout.Height);
+                rawImage.raw.dim = new Point2D((uint)bufferLayout.Width, (uint)bufferLayout.Height);
                 rawImage.Init();
                 unsafe
                 {
                     ((IMemoryBufferByteAccess)reference).GetBuffer(out var temp, out uint capacity);
                     for (int y = 0; y < rawImage.raw.dim.height; y++)
                     {
-                        int realY = y * rawImage.raw.dim.width * 3;
-                        int bufferY = y * rawImage.raw.dim.width * 4 + +bufferLayout.StartIndex;
+                        long realY = y * rawImage.raw.dim.width * 3;
+                        long bufferY = y * rawImage.raw.dim.width * 4 + +bufferLayout.StartIndex;
                         for (int x = 0; x < rawImage.raw.dim.width; x++)
                         {
-                            int realPix = realY + (3 * x);
-                            int bufferPix = bufferY + (4 * x);
+                            long realPix = realY + (3 * x);
+                            long bufferPix = bufferY + (4 * x);
                             rawImage.raw.data[realPix] = temp[bufferPix + 2];
                             rawImage.raw.data[realPix + 1] = temp[bufferPix + 1];
                             rawImage.raw.data[realPix + 2] = temp[bufferPix];
